blockchain: Introduce the `Blockchain` class
The method `Block.get_from_hash()` has been removed (*). Instead, multiple blocks
can now be stored in a dedicated `Blockchain`.

(*) Because of this, a `Block` can't be deserialized anymore. Deserialization
required to access the previous block via `Block.get_from_hash(prev_block_hash)`.
However, looking at the API of the newly introduced `Blockchain` class,
it shouldn't be necessary to deserialize a `Block` directly: A `Block` can be
reconstructed by calling `blockchain.push_head_from_json_compatible()`. It
requires as parameters

  * the serialized form of a `Block` and
  * the list of transactions included within that `Block`

and has the additional requirement that the previous `Block` is contained in the
`blockchain`. This also preserves the invariant that only valid blocks exist as
`Block` objects since `Block`s are only ever constructed via the constructor
which raises a `ValueError` if fed with an invalid transaction.